+++ src/vue/pages/sale-details/SaleOverviewDetails.vue
@@ -53,6 +53,21 @@
 
         <tr>
           <td>
+            {{ 'sale-overview.whitelisted-title' | globalize }}
+          </td>
+          <td>
+            <template v-if="sale.isWhitelisted">
+              {{ 'sale-overview.yes-msg' | globalize }}
+            </template>
+
+            <template v-else>
+              {{ 'sale-overview.no-msg' | globalize }}
+            </template>
+          </td>
+        </tr>
+
+        <tr>
+          <td>
             <!-- eslint-disable-next-line max-len -->
             {{ 'sale-overview.owner-lbl' | globalize({ asset: sale.baseAsset }) }}
           </td>